+++ app/middleware/graphers/turtle_utils.py
@@ -1,4 +1,3 @@
-import urlparse
 import config  # this is the config.py
 from datetime import datetime
 from rdflib import Namespace, URIRef, Literal
